[protoc-gen-go-jsonpb] Only emit output for requested files
As the documentation of "CodeGenerationRequest" message says, the code
generator should generate code only for the files listed in the repeated
FileToGenerate field.  Generating code for other, imported protobuf files
will result in multiple bazel targets trying to write the same files
under bazel-genfiles, which is not good! (For example, if you have a.proto
and b.proto both importing c.proto, and we want to generate .pb.jsonpb.go
files for all of them, we don't want a.proto and b.proto generating
c.pb.jsonpb.go file).